Finished ActiveEffects, implemented Fated rolls ActiveEffects now fully implemented, toggled via a function on the item class (toggleEffects). All effects are toggled at once. There are no plans to allow indidivual item effects to be toggled one by one.
FATED is now implemented as a custom die model. This die treats all 1's
as 0's, leading to double the Setbacks, to account for characters taking
the "FATED" trait.

FATED is implemented via a character flag (c2d10.fated) which is set to
true by the trait.

Finally, added localization for all effects-related areas.