Addressed inability to alter asset amount
On the character's asset tab, it was not possible to alter the amount of
a certain asset, following the universal item management updates.

This was due to the overall clickbox overriding the amount listener.
This has been addressed by removing the clickbox from the TR class
element and adding clickables to the name and type instead.